team-vanilla/requirements.txt
hata==1.1.5
attrs~=20.3.0
cattrs